pyproject.toml
[tool.poetry]
name        = "devatapath-bot"
version     = "0.1.0"
description = "Telegram-бот для мини-курсов"
authors     = ["Ваша Фамилия <you@example.com>"]

[build-system]
requires    = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.6"
aiohttp = "^3.9"
python-dotenv = "^1.1"
SQLAlchemy = "^2.0"
aiosqlite = "^0.19"